IT WORKS
Pipeline works end to end

+++ Scripts/tasks_worker.py
@@ -2,9 +2,7 @@ from celery import Celery
 import subprocess, os, shutil
 from task_shared_services import celery_url_path, file_size_mb, task_start_time, task_duration_time
 
-backend_path = celery_url_path('rpc://') 
-broker_path = celery_url_path('amqp://') 
-app = Celery('tasks', backend = backend_path, broker = broker_path)
+app = Celery('tasks', backend = celery_url_path('rpc://'), broker = celery_url_path('amqp://') )
 
 @app.task(queue='worker')
 def fencoder(ffmpeg_inputs):
@@ -16,6 +14,9 @@ def fencoder(ffmpeg_inputs):
   
     ffmpeg_command = ffmpeg_inputs['ffmpeg_command']
 
+    print ('FFMpeg command for ' + ffmpeg_inputs['file'])
+    print (ffmpeg_command)
+
     print ('Please hold')
     
     process = subprocess.Popen(ffmpeg_command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT,universal_newlines=True)
@@ -23,6 +24,7 @@ def fencoder(ffmpeg_inputs):
     for line in process.stdout:
         print(line)
 
+    files_exist = str()
     encode_outcome = str()
     stats = str()
     move = str()
@@ -32,47 +34,49 @@ def fencoder(ffmpeg_inputs):
 
     if os.path.exists(file_path and temp_filepath):
         files_exist = 'yes'
+    elif os.path.exists(temp_filepath):
+        os.remove(temp_filepath)
     else:
-        print('Something went wrong, a file is missing') 
+        print ('Issue with: ' + ffmpeg_inputs['file'])
 
     
     if files_exist == 'yes':
         old_file_size = file_size_mb(file_path)
         new_file_size = file_size_mb(temp_filepath)
         new_file_size_difference = old_file_size - new_file_size
-        print ('Old file size is: ' + str(old_file_size))
-        print ('New file size is: ' + str(new_file_size))
-        print ('Space saved on this encode: ' + str(new_file_size_difference))
+        print ('Old file size is: ' + str(old_file_size) + ' MB')
+        print ('New file size is: ' + str(new_file_size) + ' MB')
+        print ('Space saved on this encode: ' + str(new_file_size_difference) + ' MB')
         stats = 'exist'
-        if new_file_size_difference > 0:
+        if new_file_size_difference >= 0:
             encode_outcome = 'success'
         elif new_file_size_difference < 0:
             encode_outcome = 'file_larger'
-        elif new_file_size_difference == 0:
-            encode_outcome = 'error'
         else:
             encode_outcome = 'unknown_outcome'
         print ('Encode Outcome: ' + encode_outcome)
+    else:
+        print ('Issue with: ' + ffmpeg_inputs['file'])
 
-    if stats == 'exist':
-        if encode_outcome == 'error':
-            os.remove(temp_filepath)
-        else:
-            destination_filepath = root + '/' + os.path.basename(temp_filepath)
-            print('Moving ' + temp_filepath + ' to ' + destination_filepath)
-            os.remove(file_path)
-            shutil.move(temp_filepath, destination_filepath) 
-            move = 'complete'
-            print ('Move complete')
+    if stats == 'exist' and encode_outcome in ['success','file_larger']:
+        destination_filepath = root + '/' + os.path.basename(temp_filepath)
+        print('Moving ' + temp_filepath + ' to ' + destination_filepath)
+        os.remove(file_path)
+        shutil.move(temp_filepath, destination_filepath) 
+        move = 'complete'
+        print ('Move complete')
+    else:
+        print ('Issue with: ' + ffmpeg_inputs['file'])
 
     if move == 'complete':
-        ffresults = ffmpeg_inputs
+        ffresults_input = ffmpeg_inputs
         del ffmpeg_inputs['temp_filepath']
-        ffmpeg_inputs.update({'new_file_size':new_file_size})
-        ffmpeg_inputs.update({'old_file_size':old_file_size})
-        ffmpeg_inputs.update({'new_file_size_difference':new_file_size_difference})
-        ffmpeg_inputs.update({'encode_outcome':encode_outcome})
-        ffresults.delay(ffmpeg_inputs)
+        ffresults_input.update({'new_file_size':new_file_size})
+        ffresults_input.update({'old_file_size':old_file_size})
+        ffresults_input.update({'new_file_size_difference':new_file_size_difference})
+        ffresults_input.update({'encode_outcome':encode_outcome})
+
+        ffresults.delay(ffresults_input)
 
         print ('ffresults called')
